[tests] regen_drctest.sh: look for pcbtest in multiple locations

+++ tests/tools/regen_drctest.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 
-PCB=../../../pcbtest.sh ./run_tests.sh --regen DRCTests
+
+if [ -L ../pcbtest.sh ]; then
+  PCB=../../../pcbtest.sh 
+elif [ -L ../pcbtest ]; then
+  PCB=../../../pcbtest
+fi
+
+PCB=$PCB ./run_tests.sh --regen DRCTests
+
+if [ $? -ne 0 ]; then
+  exit 1
+fi
 
 rm golden/DRCTests/*-flags-after.txt
 rm golden/DRCTests/*-flags-before.txt
